correction
Correction et création d'un fichier de fonction
