+++ Raspberry/caméra/traitement d'image/.idea/workspace.xml
@@ -25,6 +25,13 @@
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
     <option name="LAST_RESOLUTION" value="IGNORE" />
   </component>
+  <component name="FileTemplateManagerImpl">
+    <option name="RECENT_TEMPLATES">
+      <list>
+        <option value="Python Script" />
+      </list>
+    </option>
+  </component>
   <component name="Git.Settings">
     <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$/../../.." />
   </component>
@@ -81,24 +88,25 @@
         </entry>
       </map>
     </option>
+    <option name="oldMeFiltersMigrated" value="true" />
   </component>
   <component name="WindowStateProjectService">
-    <state width="1899" height="348" key="GridCell.Tab.0.bottom" timestamp="1631967163715">
+    <state width="1899" height="281" key="GridCell.Tab.0.bottom" timestamp="1631976113186">
       <screen x="0" y="0" width="1920" height="1040" />
     </state>
-    <state width="1899" height="348" key="GridCell.Tab.0.bottom/0.0.1920.1040/1920.0.1360.728@0.0.1920.1040" timestamp="1631967163715" />
-    <state width="1899" height="348" key="GridCell.Tab.0.center" timestamp="1631967163715">
+    <state width="1899" height="281" key="GridCell.Tab.0.bottom/0.0.1920.1040/1920.0.1360.728@0.0.1920.1040" timestamp="1631976113186" />
+    <state width="1899" height="281" key="GridCell.Tab.0.center" timestamp="1631976113186">
       <screen x="0" y="0" width="1920" height="1040" />
     </state>
-    <state width="1899" height="348" key="GridCell.Tab.0.center/0.0.1920.1040/1920.0.1360.728@0.0.1920.1040" timestamp="1631967163715" />
-    <state width="1899" height="348" key="GridCell.Tab.0.left" timestamp="1631967163715">
+    <state width="1899" height="281" key="GridCell.Tab.0.center/0.0.1920.1040/1920.0.1360.728@0.0.1920.1040" timestamp="1631976113186" />
+    <state width="1899" height="281" key="GridCell.Tab.0.left" timestamp="1631976113185">
       <screen x="0" y="0" width="1920" height="1040" />
     </state>
-    <state width="1899" height="348" key="GridCell.Tab.0.left/0.0.1920.1040/1920.0.1360.728@0.0.1920.1040" timestamp="1631967163715" />
-    <state width="1899" height="348" key="GridCell.Tab.0.right" timestamp="1631967163715">
+    <state width="1899" height="281" key="GridCell.Tab.0.left/0.0.1920.1040/1920.0.1360.728@0.0.1920.1040" timestamp="1631976113185" />
+    <state width="1899" height="281" key="GridCell.Tab.0.right" timestamp="1631976113186">
       <screen x="0" y="0" width="1920" height="1040" />
     </state>
-    <state width="1899" height="348" key="GridCell.Tab.0.right/0.0.1920.1040/1920.0.1360.728@0.0.1920.1040" timestamp="1631967163715" />
+    <state width="1899" height="281" key="GridCell.Tab.0.right/0.0.1920.1040/1920.0.1360.728@0.0.1920.1040" timestamp="1631976113186" />
     <state width="1899" height="348" key="GridCell.Tab.1.bottom" timestamp="1631967163715">
       <screen x="0" y="0" width="1920" height="1040" />
     </state>
